Rename ray_batched_sampler to ray_sampler
