+++ src/garage/sampler/__init__.py
@@ -2,7 +2,7 @@ from garage.sampler.base import BaseSampler
 from garage.sampler.base import Sampler
 from garage.sampler.batch_sampler import BatchSampler
 from garage.sampler.is_sampler import ISSampler
-from garage.sampler.ray_batched_sampler import RaySampler, SamplerWorker
+from garage.sampler.ray_sampler import RaySampler, SamplerWorker
 from garage.sampler.stateful_pool import singleton_pool
 
 __all__ = [
